+++ lib/e_way_client/requests/query_list_provinces_request.rb
@@ -1,5 +1,14 @@
 module EWayClient
   class QueryListProvincesRequest < BaseRequest
 
+    private
+
+    def default_message
+      {
+        "USERNAME" => username,
+        "PASSWORD" => password,
+      }
+    end
+
   end
 end